pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-git-tagging"
version = "0.0.1a4"
description = "Tags git repositories"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact/git-tagging"
packages = [{include = "pythonedaartifactgittagging"}]

[tool.poetry.dependencies]
python = "^3.8"
pythoneda-base = "^0.0.1a15"
pythoneda-shared-git = "^0.0.1a1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
